src/ai/requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
pytrends>=4.9.0
python-dotenv>=0.19.0 